requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0
datasets>=2.14.0
sentence-transformers>=2.2.0

# LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0

# Vector store
chromadb>=0.4.0

# UI
streamlit>=1.29.0

# Environment
python-dotenv>=1.0.0

# PDF (for PRD reading)
PyPDF2>=3.0.0
